refactor(wallet-sdk): adjust test script & import paths
