+++ sdk/wallet-sdk/src/index.ts
@@ -32,6 +32,7 @@ export {
     Enums_ParticipantPermission,
 } from '@canton-network/core-ledger-proto'
 export * from './config.js'
+export * from './tokenStandardController.js'
 import { PartyId } from '@canton-network/core-types'
 import { AuthTokenProvider } from './authTokenProvider.js'
 export { AuthTokenProvider } from './authTokenProvider.js'
