+++ sdk/wallet-sdk/src/v1/logger/logger.ts
@@ -1,8 +1,8 @@
 // Copyright (c) 2025-2026 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import ConsoleLogAdapter from './adapter/console'
-import PinoLogAdapter from './adapter/pino'
+import ConsoleLogAdapter from './adapter/console.js'
+import PinoLogAdapter from './adapter/pino.js'
 import {
     AllowedLogAdapters,
     LogAdapter,
